# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Infinity Knowledge Control Plane API
  version: 1.0.0
  description: |
    Machine-readable contract for the implemented `/api/v1` HTTP surface.
    Every operation requires an OIDC bearer token. Operations described as
    administrator-only additionally require the `knowledge-admin` realm role.
    Tenant and principal identity are derived exclusively from the validated token.
servers:
  - url: http://localhost:8080
    description: Local control plane
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/documents/files:
    post:
      tags: [Documents]
      operationId: ingestFile
      summary: Retain, parse and ingest a supported source file
      description: Requires `knowledge-admin`.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/FileDocumentRequest'
      responses:
        '200':
          description: Active revision and retained source metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDocumentResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          $ref: '#/components/responses/SourceTooLarge'
  /api/v1/documents/{documentId}/source:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      tags: [Documents]
      operationId: getDocumentSourceMetadata
      summary: Get authorized original-source metadata
      description: Tenant administrators may inspect retained inactive sources; other callers must have access to an active document.
      responses:
        '200':
          description: Source metadata without object-store coordinates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceObjectResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/DocumentNotFound'
  /api/v1/documents/{documentId}/source/content:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      tags: [Documents]
      operationId: downloadDocumentSource
      summary: Stream an authorized original source
      description: Tenant administrators may inspect retained inactive sources; other callers must have access to an active document.
      parameters:
        - name: inline
          in: query
          required: false
          description: Request a sandboxed inline response when the media type is previewable.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Source bytes; unsafe types remain attachments even when inline is requested
          headers:
            Content-Disposition:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            text/plain:
              schema:
                type: string
                format: binary
            text/markdown:
              schema:
                type: string
                format: binary
            text/html:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/DocumentNotFound'
...
  /api/v1/evaluations/datasets:
    get:
      tags: [Evaluations]
      operationId: listEvaluationDatasets
      summary: List tenant evaluation datasets
      description: Requires `knowledge-admin`.
      responses:
        '200':
          description: Evaluation datasets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EvaluationDataset'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Evaluations]
      operationId: createEvaluationDataset
      summary: Create an evaluation dataset
      description: Requires `knowledge-admin`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEvaluationDatasetRequest'
      responses:
        '201':
          description: Created dataset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationDataset'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/evaluations/datasets/{datasetId}/cases:
    parameters:
      - $ref: '#/components/parameters/DatasetId'
    get:
      tags: [Evaluations]
      operationId: listEvaluationCases
      summary: List cases in an evaluation dataset
      description: Requires `knowledge-admin`.
      responses:
        '200':
          description: Evaluation cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EvaluationCase'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Evaluations]
      operationId: createEvaluationCase
      summary: Add a labeled case to an evaluation dataset
      description: Requires `knowledge-admin`; at least one expected document or chunk is required.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEvaluationCaseRequest'
      responses:
        '201':
          description: Created case
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationCase'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/evaluations/datasets/{datasetId}/runs:
    parameters:
      - $ref: '#/components/parameters/DatasetId'
    get:
      tags: [Evaluations]
      operationId: listEvaluationRuns
      summary: List runs in an evaluation dataset
      description: Requires `knowledge-admin`; list entries omit case results.
      responses:
        '200':
          description: Evaluation runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EvaluationRun'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Evaluations]
      operationId: startEvaluationRun
      summary: Start an asynchronous evaluation run
      description: Requires `knowledge-admin`; the dataset must contain at least one case.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartEvaluationRunRequest'
      responses:
        '202':
          description: Accepted evaluation run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationRun'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/OperationInProgress'
        '503':
          $ref: '#/components/responses/WorkQueueSaturated'
  /api/v1/evaluations/datasets/{datasetId}/compare:
    parameters:
      - $ref: '#/components/parameters/DatasetId'
    post:
      tags: [Evaluations]
      operationId: compareEvaluationRuns
      summary: Compare completed candidate and baseline runs
      description: Requires `knowledge-admin`; both distinct runs must belong to this dataset and be `SUCCEEDED`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompareEvaluationRunsRequest'
      responses:
        '200':
          description: Quality-gate comparison
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationRunComparison'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/evaluations/runs/{runId}:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      tags: [Evaluations]
      operationId: getEvaluationRun
      summary: Get one evaluation run with case results
      description: Requires `knowledge-admin`.
      responses:
        '200':
          description: Evaluation run detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationRun'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    DatasetId:
      name: datasetId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Bearer token is absent or invalid
    Forbidden:
      description: The authenticated principal is not allowed to perform the operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InvalidRequest:
      description: Request validation, parsing or domain argument failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    DocumentNotFound:
      description: The requested authorized document source does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    OperationInProgress:
      description: The same resource already has an operation in progress
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    WorkQueueSaturated:
      description: The bounded asynchronous work queue cannot accept more work
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    SourceTooLarge:
      description: Uploaded source exceeds the configured limit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiError:
      type: object
      required: [code, message, requestId, timestamp]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
...
    FileDocumentRequest:
      type: object
      required: [file, spaceId, externalId]
      properties:
        file:
          type: string
          format: binary
        spaceId:
          type: string
          minLength: 1
          maxLength: 64
        externalId:
          type: string
          minLength: 1
          maxLength: 512
        title:
          type: string
          maxLength: 512
        language:
          type: string
          minLength: 1
          maxLength: 32
          default: zh-CN
        authority:
          type: integer
          minimum: 0
          maximum: 100
          default: 80
    FileDocumentResponse:
      type: object
      required: [documentId, revisionId, changed, elementCount, chunkCount, originalFileName, mediaType, contentLength, vectorStatus, warnings]
      properties:
        documentId:
          type: string
          format: uuid
        revisionId:
          type: string
          format: uuid
        changed:
          type: boolean
        elementCount:
          type: integer
        chunkCount:
          type: integer
        originalFileName:
          type: string
        mediaType:
          type: string
        contentLength:
          type: integer
          format: int64
        vectorStatus:
          type: string
        warnings:
          type: array
          items:
            type: string
    SourceObjectResponse:
      type: object
      required: [documentId, revisionId, documentStatus, originalFileName, mediaType, contentLength, checksumSha256, storedAt, previewable]
      properties:
        documentId:
          type: string
          format: uuid
        revisionId:
          type: string
          format: uuid
        documentStatus:
          $ref: '#/components/schemas/DocumentStatus'
        originalFileName:
          type: string
        mediaType:
          type: string
        contentLength:
          type: integer
          format: int64
        checksumSha256:
          type: string
        storedAt:
          type: string
          format: date-time
        previewable:
          type: boolean
...
    DocumentStatus:
      type: string
      enum: [DRAFT, ACTIVE, DEPRECATED, ARCHIVED, DELETED]
...
    CreateEvaluationDatasetRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 256
        description:
          type: string
          maxLength: 4000
    CreateEvaluationCaseRequest:
      type: object
      required: [query, spaceIds]
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 16000
        spaceIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            minLength: 1
        expectedDocuments:
          type: array
          uniqueItems: true
          default: []
          items:
            type: string
            format: uuid
        expectedChunks:
          type: array
          uniqueItems: true
          default: []
          items:
            type: string
            format: uuid
        topK:
          type: integer
          minimum: 1
          maximum: 100
        labels:
          type: object
          default: {}
          additionalProperties:
            type: string
    StartEvaluationRunRequest:
      type: object
      properties:
        topK:
          type: integer
          minimum: 1
          maximum: 100
        configuration:
          type: object
          default: {}
          additionalProperties:
            type: string
    CompareEvaluationRunsRequest:
      type: object
      required: [baselineRunId, candidateRunId]
      properties:
        baselineRunId:
          type: string
          format: uuid
        candidateRunId:
          type: string
          format: uuid
        minimumHitRate:
          $ref: '#/components/schemas/UnitInterval'
        minimumRecallAtK:
          $ref: '#/components/schemas/UnitInterval'
        minimumMrr:
          $ref: '#/components/schemas/UnitInterval'
        minimumNdcgAtK:
          $ref: '#/components/schemas/UnitInterval'
        maximumRegression:
          allOf:
            - $ref: '#/components/schemas/UnitInterval'
          default: 0
    UnitInterval:
      type: number
      format: double
      minimum: 0
      maximum: 1
      nullable: true
    EvaluationDataset:
      type: object
      required: [id, name, description, version, status, caseCount, runCount, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        version:
          type: integer
          format: int64
        status:
          type: string
        caseCount:
          type: integer
          format: int64
        runCount:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
    EvaluationCase:
      type: object
      required: [id, datasetId, query, spaceIds, expectedDocuments, expectedChunks, topK, labels, createdAt]
      properties:
        id:
          type: string
          format: uuid
        datasetId:
          type: string
          format: uuid
        query:
          type: string
        spaceIds:
          type: array
          uniqueItems: true
          items:
            type: string
        expectedDocuments:
          type: array
          uniqueItems: true
          items:
            type: string
            format: uuid
        expectedChunks:
          type: array
          uniqueItems: true
          items:
            type: string
            format: uuid
        topK:
          type: integer
        labels:
          type: object
          additionalProperties:
            type: string
        createdAt:
          type: string
          format: date-time
    EvaluationRun:
      type: object
      required: [id, datasetId, status, caseCount, failedCaseCount, configuration, metrics, requestedBy, errorCode, startedAt, completedAt, results]
      properties:
        id:
          type: string
          format: uuid
        datasetId:
          type: string
          format: uuid
        status:
          type: string
        caseCount:
          type: integer
        failedCaseCount:
          type: integer
        configuration:
          type: object
          additionalProperties: true
        metrics:
          type: object
          additionalProperties: true
        requestedBy:
          type: string
        errorCode:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/EvaluationCaseResult'
    EvaluationCaseResult:
      type: object
      required: [caseId, traceId, status, hit, recallAtK, reciprocalRank, ndcgAtK, resultCount, durationMs, errorCode]
      properties:
        caseId:
          type: string
          format: uuid
        traceId:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
        hit:
          type: boolean
        recallAtK:
          type: number
          format: double
        reciprocalRank:
          type: number
          format: double
        ndcgAtK:
          type: number
          format: double
        resultCount:
          type: integer
        durationMs:
          type: integer
          format: int64
        errorCode:
          type: string
          nullable: true
    EvaluationRunComparison:
      type: object
      required: [datasetId, baselineRunId, candidateRunId, baseline, candidate, deltas, maximumRegression, passed, violations]
      properties:
        datasetId:
          type: string
          format: uuid
        baselineRunId:
          type: string
          format: uuid
        candidateRunId:
          type: string
          format: uuid
        baseline:
          $ref: '#/components/schemas/MetricMap'
        candidate:
          $ref: '#/components/schemas/MetricMap'
        deltas:
          $ref: '#/components/schemas/MetricMap'
        maximumRegression:
          type: number
          format: double
        passed:
          type: boolean
        violations:
          type: array
          items:
            $ref: '#/components/schemas/EvaluationGateViolation'
    MetricMap:
      type: object
      additionalProperties:
        type: number
        format: double
    EvaluationGateViolation:
      type: object
      required: [metric, rule, expected, actual]
      properties:
        metric:
          type: string
        rule:
          type: string
          enum: [MINIMUM, MAXIMUM_REGRESSION]
        expected:
          type: number
          format: double
        actual:
          type: number
          format: double
...